refactor(explore): keep Prometheus types out of shared explore state
Carry the search strategy's data frame meta through the result cache generically and
keep the language query options declaration free of PromQL fields, so the Prometheus
option and step resolution types are only referenced under pages/metrics.
