+++ src/plugins/explore/public/application/utils/languages/types.ts
@@ -4,10 +4,14 @@
  */
 
 import { Query } from '../../../../../data/common';
-import type { PromQLQueryOptions } from '../../../../../query_enhancements/common';
 
-export type { PerQueryOptions, PromQLQueryOptions } from '../../../../../query_enhancements/common';
+export interface LanguageQueryOptions {
+  maxDataPoints?: number;
+  /** One entry per query segment, in segment order. The entry shape belongs to the language,
+   * so narrow it there rather than naming a language's fields here. */
+  perQueryOptions?: unknown[];
+}
 
 export type QueryWithQueryAsString = Omit<Query, 'query'> & {
   query: string;
-} & PromQLQueryOptions;
+} & LanguageQueryOptions;
