+++ src/plugins/explore/public/application/utils/state_management/actions/query_actions.ts
@@ -704,7 +704,7 @@ const executeQueryBase = async (
       elapsedMs: inspectorRequest.getTime()!,
       fieldSchema: searchSource.getDataFrame()?.schema,
       profile: searchSource.getDataFrame()?.meta?.profile,
-      stepResolution: searchSource.getDataFrame()?.meta?.stepResolution,
+      frameMeta: searchSource.getDataFrame()?.meta,
     };
 
     if (isHistogramQuery && effectiveHistogramConfig) {
